test(gui): split the Scrolling caption suite off the placement one
`LayoutSchematicScrollingTests` reached 317 lines holding both
the drawing and the words beside it, and tests.md asks for the
split before the ceiling rather than at it. They fail apart:
one holds where the anchor rests the focus and where the engine
opens the window, the other holds what the caption may say about
either.
