+++ Tests/KiwiDeskGuiTests/LayoutSchematicScrollingTests.swift
@@ -27,6 +27,9 @@ import Testing
 /// rather than a second copy, because a copy of the rule under
 /// guard leaves one suite green on the retired rule.
 ///
+/// The words beside the frame are `LayoutSchematicCaptionTests`',
+/// split off before this file reached the ceiling.
+///
 /// `@MainActor` because the derived quantities are properties of
 /// `View`s, which are main-actor isolated: off the main actor the
 /// first read traps in the concurrency runtime rather than failing
@@ -137,123 +140,11 @@ struct LayoutSchematicScrollingTests {
         }
     }
 
-    /// The words switch on the anchor, because the anchor is the
-    /// only thing `follow` changes: its frame is `center`'s to the
-    /// pixel (the test above), so a caption shared with `center`
-    /// makes selecting Follow a no-op on screen — and made
-    /// VoiceOver assert Follow's pan over a tile drawn with Center
-    /// selected.
-    @Test("Follow says something Center does not")
-    func theAnchorReachesTheWords() {
-        let follow = scrolling(anchor: .follow, scale: .panel)
-        let centre = scrolling(anchor: .center, scale: .panel)
-        #expect(follow.caption != centre.caption)
-        #expect(follow.axLabel != centre.axLabel)
-        // And the anchored three are not carrying a sentence about
-        // a fourth: Follow's own name appears in Follow's words
-        // and nowhere else.
-        let name = L("scroll_grid.anchor.follow", "Follow")
-        #expect(follow.caption.contains(name))
-        for anchor in [
-            ScrollingParams.Anchor.center, .start, .end,
-        ] {
-            let other = scrolling(anchor: anchor, scale: .panel)
-            #expect(!other.caption.contains(name))
-            #expect(!other.axLabel.contains(name))
-        }
-    }
-
-    /// The caption promises a `+` only where one is drawn.
-    ///
-    /// The row runs several canvases wide at most counts, so the
-    /// incoming slot is often clipped away entirely — at the
-    /// default five windows with New window ▸ Last it already is.
-    /// The clause is a condition on the row rather than on the
-    /// pixels, since the caption cannot read the canvas, so what
-    /// is owed is the IMPLICATION: wherever it is claimed, the
-    /// drawing's own `onCanvas` agrees, at every width a pane can
-    /// be. Both directions of the pair are exercised below, so a
-    /// clause that never fires cannot pass for a correct one.
-    @Test("the + clause is claimed only where the + is drawn")
-    func theInsertionClauseMatchesTheDrawing() {
-        var claimed = 0
-        var withheld = 0
-        for anchor in ScrollingParams.Anchor.allCases {
-            for placement in placements {
-                for size in slotSizes {
-                    for count in LayoutSchematic.windowCountRange {
-                        let schematic = scrolling(
-                            anchor: anchor,
-                            placement: placement,
-                            slotSize: size,
-                            windows: count,
-                            scale: .panel
-                        )
-                        if schematic.drawsInsertionMark {
-                            claimed += 1
-                        } else {
-                            withheld += 1
-                            continue
-                        }
-                        for along in widths {
-                            let m = schematic.metrics(along: along)
-                            #expect(
-                                schematic.onCanvas(
-                                    m.newIdx,
-                                    m,
-                                    along: along
-                                ),
-                                Comment(
-                                    rawValue:
-                                        "\(anchor)/\(placement) "
-                                        + "at \(count), \(along) "
-                                        + "pt: + off the canvas"
-                                )
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        #expect(claimed > 0)
-        #expect(withheld > 0)
-        // And the clause is what the claim renders: the sentence
-        // is present exactly when the mark is.
-        let beside = scrolling(placement: .afterFocused, scale: .panel)
-        let far = scrolling(
-            placement: .last,
-            windows: LayoutSchematic.windowCountRange.upperBound,
-            scale: .panel
-        )
-        #expect(beside.drawsInsertionMark)
-        #expect(!far.drawsInsertionMark)
-        #expect(!beside.insertionClause.isEmpty)
-        #expect(far.insertionClause.isEmpty)
-        #expect(beside.caption.contains(beside.insertionClause))
-        #expect(far.caption.count < beside.caption.count)
-        // No dangling space where the clause went.
-        #expect(!far.caption.hasSuffix(" "))
-    }
-
     // MARK: - The promise, and the fixtures
 
     private let placements: [SpawnPlacement] = [
         .first, .last, .beforeFocused, .afterFocused,
     ]
-
-    /// The slot sizes the three `ScrollSize` arms can produce,
-    /// including both ends of the points ramp: the 14 pt floor and
-    /// the widest slot are where a width-free claim would break if
-    /// it were going to.
-    private let slotSizes: [ScrollSize] = [
-        .auto, .fraction(0.12), .fraction(0.92), .points(100),
-        .points(1200),
-    ]
-
-    /// The along-axis lengths a panel can take, from a pane
-    /// squeezed beside the override rows at the 720 pt minimum
-    /// window to a full-screen one.
-    private let widths: [CGFloat] = [60, 120, 228, 400, 900, 1600]
 
     /// Where each anchor rests the focused window, stated
     /// independently of the schematic's own switch.
@@ -301,17 +192,14 @@ struct LayoutSchematicScrollingTests {
         orientation: ScrollingParams.Orientation = .horizontal,
         anchor: ScrollingParams.Anchor = .center,
         placement: SpawnPlacement = .last,
-        slotSize: ScrollSize = .auto,
-        windows: Int = LayoutSchematic.defaultWindowCount,
-        scale: SchematicScale = .tile
+        windows: Int = LayoutSchematic.defaultWindowCount
     ) -> ScrollingSchematic {
         ScrollingSchematic(
             orientation: orientation,
             anchor: anchor,
-            slotSize: slotSize,
+            slotSize: .auto,
             placement: placement,
-            windows: windows,
-            scale: scale
+            windows: windows
         )
     }
 }
